race detection and log generation now work
to force execution of syscall B before syscall A, we now
find a crosscut and generate a bookmark with negative syscnts,
and another similar bookmark except for syscall B that gets
a positive bookmark. between these two bookmarks for syscall B
we inject a diable-resource-lock flag change so that the syscall
won't cause divergence.

* update logic for recent scribe "shared-read" logging
* add injection of "modify scrine_ps->flags"
* refactor and cleanup save_events()
* minor (but important!) fixes to crosscut_graph
* make show_races() work